+++ CHANGELOG.md
@@ -16,6 +16,11 @@
  - Added `controller_name_postfix` config option to allow the user to change the controller post-fix or even remove it altogether.
  - Added `form_request_name_postfix` config option to allow the user to change the form-request post-fix or even remove it altogether.
  - You can use Laravel 5.5 custom validation rule directly in the validations string.
+ -- Added `create_move_file_method` config option to allow the user to chose not to create moveFile method on every CRUD when file-upload is required.
+
+## Template Changes
+ - A new stub was added controller-getdata-method-5.5.stub to allow you to simplify your validation when using Laravel 5.5! This will make your code much cleaner and simpler thanks to the new Laravel 5.5 validation.
+ - The stub `controller-getdata-method.stub` has a slight change to increase the security of the generate code. instead of using `$request->all()` to get all the data from the request, we do `$request->only([...])` to only get the fields that should be needed only!